ai-interview/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.10
alembic>=1.12.0
redis>=5.0.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
reportlab>=4.0.0
pdfminer.six>=20221105
sentence-transformers>=2.2.0
faiss-cpu>=1.9.0
numpy>=1.24.0
email-validator>=2.1.0
python-dotenv>=1.0.0